[Refactor] BankAccountUpperView로 상단 section 뷰 분리
